# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Personal Health System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /logs:
    get:
      operationId: getLogs
      tags: [logs]
      summary: List daily logs
      parameters:
        - name: from
          in: query
          schema:
            type: string
            format: date
        - name: to
          in: query
          schema:
            type: string
            format: date
        - name: limit
          in: query
          schema:
            type: integer
            default: 30
      responses:
        "200":
          description: List of daily logs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/DailyLog"
  /logs/{date}:
    get:
      operationId: getDailyLog
      tags: [logs]
      summary: Get daily log by date
      parameters:
        - name: date
          in: path
          required: true
          schema:
            type: string
            format: date
      responses:
        "200":
          description: Daily log
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DailyLog"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: patchDailyLog
      tags: [logs]
      summary: Patch daily log fields (water, sleep, mood, etc.)
      parameters:
        - name: date
          in: path
          required: true
          schema:
            type: string
            format: date
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PatchDailyLogBody"
      responses:
        "200":
          description: Updated daily log
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DailyLog"
  /logs/{date}/food:
    post:
      operationId: addFoodEntry
      tags: [logs]
      summary: Add food entry to a daily log
      parameters:
        - name: date
          in: path
          required: true
          schema:
            type: string
            format: date
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddFoodEntryBody"
      responses:
        "201":
          description: Food entry added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FoodEntry"
  /logs/{date}/food/{entryId}:
    delete:
      operationId: deleteFoodEntry
      tags: [logs]
      summary: Delete a food entry from a daily log
      parameters:
        - name: date
          in: path
          required: true
          schema:
            type: string
            format: date
        - name: entryId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /logs/{date}/activity:
    post:
      operationId: addActivityEntry
      tags: [logs]
      summary: Add activity entry to a daily log
      parameters:
        - name: date
          in: path
          required: true
          schema:
            type: string
            format: date
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddActivityEntryBody"
      responses:
        "201":
          description: Activity entry added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ActivityEntry"
  /logs/{date}/activity/{entryId}:
    delete:
      operationId: deleteActivityEntry
      tags: [logs]
      summary: Delete an activity entry
      parameters:
        - name: date
          in: path
          required: true
          schema:
            type: string
            format: date
        - name: entryId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /labs:
    get:
      operationId: getLabs
      tags: [labs]
      summary: List lab results
      responses:
        "200":
          description: Lab results
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/LabResult"
    post:
      operationId: addLabResult
      tags: [labs]
      summary: Add a lab result
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddLabResultBody"
      responses:
        "201":
          description: Lab result created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LabResult"
  /labs/{id}:
    delete:
      operationId: deleteLabResult
      tags: [labs]
      summary: Delete a lab result
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
      required: [success]
...
    NutrientData:
      type: object
      description: Complete nutrient breakdown of a food portion
      properties:
        calories:
          type: number
        protein:
          type: number
        fat:
          type: number
        carbs:
          type: number
        fiber:
          type: number
        sugar:
          type: number
        sodium:
          type: number
        potassium:
          type: number
        calcium:
          type: number
        magnesium:
          type: number
        iron:
          type: number
        zinc:
          type: number
        vitaminA:
          type: number
        vitaminC:
          type: number
        vitaminD:
          type: number
        vitaminE:
          type: number
        vitaminK:
          type: number
        vitaminB1:
          type: number
        vitaminB2:
          type: number
        vitaminB3:
          type: number
        vitaminB6:
          type: number
        vitaminB12:
          type: number
        folate:
          type: number
        omega3:
          type: number
        omega6:
          type: number
        saturatedFat:
          type: number
        cholesterol:
          type: number
      required: [calories, protein, fat, carbs]
...
    FoodEntry:
      type: object
      properties:
        id:
          type: integer
        date:
          type: string
          format: date
        foodId:
          type: string
          nullable: true
        foodName:
          type: string
        mealType:
          type: string
          enum: [breakfast, lunch, dinner, snack]
        amount:
          type: number
          description: Amount in grams
        nutrients:
          $ref: "#/components/schemas/NutrientData"
        createdAt:
          type: string
          format: date-time
      required: [id, date, foodName, mealType, amount, nutrients, createdAt]
    AddFoodEntryBody:
      type: object
      properties:
        foodId:
          type: string
          nullable: true
        foodName:
          type: string
        mealType:
          type: string
          enum: [breakfast, lunch, dinner, snack]
        amount:
          type: number
        nutrients:
          $ref: "#/components/schemas/NutrientData"
      required: [foodName, mealType, amount, nutrients]
    ActivityEntry:
      type: object
      properties:
        id:
          type: integer
        date:
          type: string
          format: date
        type:
          type: string
        category:
          type: string
          enum: [cardio, strength, flexibility, sports, other]
        durationMinutes:
          type: integer
        intensity:
          type: string
          enum: [light, moderate, vigorous]
        caloriesBurned:
          type: number
          nullable: true
        steps:
          type: integer
          nullable: true
        notes:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
      required: [id, date, type, category, durationMinutes, intensity, createdAt]
    AddActivityEntryBody:
      type: object
      properties:
        type:
          type: string
        category:
          type: string
          enum: [cardio, strength, flexibility, sports, other]
        durationMinutes:
          type: integer
        intensity:
          type: string
          enum: [light, moderate, vigorous]
        caloriesBurned:
          type: number
          nullable: true
        steps:
          type: integer
          nullable: true
        notes:
          type: string
          nullable: true
      required: [type, category, durationMinutes, intensity]
    SleepData:
      type: object
      properties:
        bedtime:
          type: string
          nullable: true
        wakeTime:
          type: string
          nullable: true
        durationHours:
          type: number
          nullable: true
        qualityScore:
          type: integer
          nullable: true
          description: 1-10 subjective quality
        wakeUps:
          type: integer
          nullable: true
        morningFeeling:
          type: string
          enum: [refreshed, okay, tired, exhausted]
          nullable: true
        notes:
          type: string
          nullable: true
    WellbeingData:
      type: object
      properties:
        energyLevel:
          type: integer
          nullable: true
          description: 1-10
        moodScore:
          type: integer
          nullable: true
          description: 1-10
        focusScore:
          type: integer
          nullable: true
          description: 1-10
        stressLevel:
          type: integer
          nullable: true
          description: 1-10
        hungerLevel:
          type: integer
          nullable: true
          description: 1-10
        symptoms:
          type: array
          items:
            type: string
        notes:
          type: string
          nullable: true
    DailyLog:
      type: object
      properties:
        id:
          type: integer
        date:
          type: string
          format: date
        waterMl:
          type: number
          description: Total water consumed in mL
        waterGoalMl:
          type: number
          nullable: true
        sleep:
          $ref: "#/components/schemas/SleepData"
          nullable: true
        wellbeing:
          $ref: "#/components/schemas/WellbeingData"
          nullable: true
        foodEntries:
          type: array
          items:
            $ref: "#/components/schemas/FoodEntry"
        activityEntries:
          type: array
          items:
            $ref: "#/components/schemas/ActivityEntry"
        totalNutrients:
          $ref: "#/components/schemas/NutrientData"
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required: [id, date, waterMl, foodEntries, activityEntries, totalNutrients, createdAt, updatedAt]
    PatchDailyLogBody:
      type: object
      properties:
        waterMl:
          type: number
          nullable: true
        waterGoalMl:
          type: number
          nullable: true
        sleep:
          $ref: "#/components/schemas/SleepData"
          nullable: true
        wellbeing:
          $ref: "#/components/schemas/WellbeingData"
          nullable: true
    LabResult:
      type: object
      properties:
        id:
          type: integer
        date:
          type: string
          format: date
        marker:
          type: string
          description: Lab marker name (e.g., ferritin, vitamin_d, hemoglobin)
        value:
          type: number
        unit:
          type: string
        refMin:
          type: number
          nullable: true
        refMax:
          type: number
          nullable: true
        laboratory:
          type: string
          nullable: true
        status:
          type: string
          enum: [normal, low, high, critical_low, critical_high, unknown]
        notes:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
      required: [id, date, marker, value, unit, status, createdAt]
    AddLabResultBody:
      type: object
      properties:
        date:
          type: string
          format: date
        marker:
          type: string
        value:
          type: number
        unit:
          type: string
        refMin:
          type: number
          nullable: true
        refMax:
          type: number
          nullable: true
        laboratory:
          type: string
          nullable: true
        notes:
          type: string
          nullable: true
      required: [date, marker, value, unit]
...